move withPlugins into config-plugins
