+++ packages/config-plugins/src/Plugin.types.ts
@@ -57,6 +57,8 @@ export interface ExportedConfigWithProps<Data = any> extends ExpoConfig {
 
 export type ConfigPlugin<Props = void> = (config: ExpoConfig, props: Props) => ExpoConfig;
 
+export type StaticPlugin<T = any> = [string | ConfigPlugin<T>, T];
+
 export type Mod<Props = any> = (
   config: ExportedConfigWithProps<Props>
 ) => OptionalPromise<ExportedConfigWithProps<Props>>;
