+++ packages/config-plugins/src/index.ts
@@ -43,4 +43,8 @@ export {
   withSettingsGradle,
 } from './plugins/android-plugins';
 
+export { withStaticPlugin } from './plugins/static-plugins';
+
 export { compileModsAsync } from './plugins/mod-compiler';
+
+export { PluginError } from './utils/errors';
