+++ packages/config-plugins/src/utils/__tests__/fixtures/project-files/my-plugin.js
