+++ packages/config-plugins/src/utils/__tests__/plugin-resolver-test.ts
@@ -1,4 +1,4 @@
-import { moduleNameIsDirectFileReference } from '../modulePluginResolver';
+import { moduleNameIsDirectFileReference } from '../plugin-resolver';
 
 describe(moduleNameIsDirectFileReference, () => {
   it('file path', () => {
