+++ packages/config/package.json
@@ -35,8 +35,8 @@
     "@babel/core": "7.9.0",
     "@expo/babel-preset-cli": "0.2.18",
     "@expo/config-types": "^40.0.0-beta.2",
+    "@expo/config-plugins": "~1.0.13",
     "@expo/json-file": "8.2.25",
-    "find-up": "^5.0.0",
     "fs-extra": "9.0.0",
     "getenv": "0.7.0",
     "glob": "7.1.6",
@@ -46,7 +46,6 @@
     "slugify": "^1.3.4"
   },
   "devDependencies": {
-    "@types/find-up": "^4.0.0",
     "@types/fs-extra": "^9.0.1",
     "memfs": "^2.15.5",
     "rimraf": "^3.0.2"
